test(gateway): extract transport unit tests
